+++ src/components/dice.js
@@ -1,9 +1,9 @@
-import logo1 from "../../assets/dice/1.png";
-import logo2 from "../../assets/dice/2.png";
-import logo3 from "../../assets/dice/3.png";
-import logo4 from "../../assets/dice/4.png";
-import logo5 from "../../assets/dice/5.png";
-import logo6 from "../../assets/dice/6.png";
+import logo1 from "../assets/dice/1.png";
+import logo2 from "../assets/dice/2.png";
+import logo3 from "../assets/dice/3.png";
+import logo4 from "../assets/dice/4.png";
+import logo5 from "../assets/dice/5.png";
+import logo6 from "../assets/dice/6.png";
 
 AFRAME.registerComponent("dice", {
   schema: {
