+++ src/scene.html
@@ -37,7 +37,7 @@
         visible="false"
       ></a-box>
       <a-entity id="probabilty-scene" probability></a-entity>
-      
+      <a-entity id="dice-manager" dice-manager></a-entity>
       <!-- Cursor -->
       <a-entity camera look-controls position="0 1.8 0">
         <a-entity
